+++ lib/pages/cad_meu_animal/take_second_picture.dart
@@ -6,26 +6,25 @@ import 'package:flutter/material.dart';
 import 'package:path/path.dart' show join;
 import 'package:path_provider/path_provider.dart';
 
-import 'package:petsaojoao/pages/cad_meu_animal/tela_confirmacao.dart';
-import 'package:petsaojoao/components/comp_cad_meu_animal/fotos_info.dart';
+import 'package:petsaojoao/components/comp_cad_meu_animal/picture_info.dart';
+import 'package:petsaojoao/components/comp_cad_meu_animal/camera_info.dart';
+import 'package:petsaojoao/pages/cad_meu_animal/take_third_picture.dart';
 
-class TirarTerceiraFoto extends StatefulWidget {
+class TirarSegundaFoto extends StatefulWidget {
   final CameraDescription camera;
   final String image1;
-  final String image2;
 
-  const TirarTerceiraFoto({
+  const TirarSegundaFoto({
     Key key,
     @required this.camera,
     @required this.image1,
-    @required this.image2,
   }) : super(key: key);
 
   @override
-  TirarTerceiraFotoState createState() => TirarTerceiraFotoState();
+  TirarSegundaFotoState createState() => TirarSegundaFotoState();
 }
 
-class TirarTerceiraFotoState extends State<TirarTerceiraFoto> {
+class TirarSegundaFotoState extends State<TirarSegundaFoto> {
   CameraController _controller;
   Future<void> _initializeControllerFuture;
 
@@ -42,7 +41,7 @@ class TirarTerceiraFotoState extends State<TirarTerceiraFoto> {
 
   @override
   // TODO: implement widget
-  TirarTerceiraFoto get widget => super.widget;
+  TirarSegundaFoto get widget => super.widget;
 
   @override
   void dispose() {
@@ -53,7 +52,6 @@ class TirarTerceiraFotoState extends State<TirarTerceiraFoto> {
   @override
   Widget build(BuildContext context) {
     String imagem1 = widget.image1;
-    String imagem2 = widget.image2;
     return Scaffold(
       body: FutureBuilder<void>(
         future: _initializeControllerFuture,
@@ -64,15 +62,7 @@ class TirarTerceiraFotoState extends State<TirarTerceiraFoto> {
                 Container(
                     height: MediaQuery.of(context).size.height / 1.5,
                     child: CameraPreview(_controller)),
-                Center(
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      Container(height: 200, child: Image.file(File(imagem1))),
-                      Container(height: 200, child: Image.file(File(imagem2)))
-                    ],
-                  ),
-                )
+                Container(height: 200, child: Image.file(File(imagem1)))
               ],
             );
           } else {
@@ -81,32 +71,31 @@ class TirarTerceiraFotoState extends State<TirarTerceiraFoto> {
         },
       ),
       floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.arrow_forward),
+        child: Icon(Icons.camera_alt),
         onPressed: () async {
           try {
             await _initializeControllerFuture;
 
             final path =
-                join((await getTemporaryDirectory()).path, 'PET-img3.png');
+                join((await getTemporaryDirectory()).path, 'PET-img2.png');
 
             if (File(path).existsSync()) {
               File(path).deleteSync(recursive: true);
               PaintingBinding.instance.imageCache.clear();
             }
 
-            await _controller.takePicture(path);
-
+            final camera = await getCameraInfo();
             final imagem1 = await getPrimeiraFoto();
             final imagem2 = await getSegundaFoto();
-            final imagem3 = await getTerceiraFoto();
+            await _controller.takePicture(path);
 
             Navigator.push(
               context,
               MaterialPageRoute(
-                builder: (context) => TelaConfirmacao(
+                builder: (context) => TirarTerceiraFoto(
+                  camera: camera,
                   image1: imagem1,
                   image2: imagem2,
-                  image3: imagem3,
                 ),
               ),
             );
