+++ lib/pages/dashboard/dashboard.dart
@@ -2,7 +2,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/material.dart';
 import 'package:petsaojoao/pages/cad_tutor/cadastro_tutor.dart';
 import 'package:petsaojoao/pages/cad_animal_encotrado/cadastro_animal_encontrado.dart';
-import 'package:petsaojoao/pages/cad_meu_animal/cad_meu_animal.dart';
+import 'package:petsaojoao/pages/cad_meu_animal/cad_my_pet.dart';
 
 //Acompanhe desing do projeto aqui --> https://www.figma.com/file/GYFrt79mzIbOUXXmFyDgwL/Material-Baseline-Design-Kit?node-id=38%3A5814
